Zombies, Barricades, Pathfinding
Zombies will move towards the player, ignoring barricades. If in range of a barricade or player they will attack them.
Once a barricade is damaged enough it will be passable for the zombie but not the player.
The player can approach a broken barricade and press F to repair it.